Add files via upload
In this file, most of the code is commented out, it all depends on how many vCenter you have. Also, I do not download the certificate for work, therefore I use the Smart Connect No SSL option(you can optionally change it if you use the certificate). Since this is python and the Pyvmomi SDK used, the running time depends on the platform and CPU used(on the ADM Ryzen 3600, the total running time was 30 seconds, on the intel core i7-8565U it was generally 1 minute). For faster work, I also used parallel polling.
P. S: I'm still new to python:) You can write to me for discussions, optimization and adding new features. There is an idea to use numba to speed up:)Thanks